// src/modules/auth/schemas/schemas.ts
import * as z from "zod";

export const SignInSchema = z.object({
    email: z.string().email({
        message: "Please enter a valid email address"
    }),
    password: z.string().min(1, {
        message: "Password is required"
    })
})

export const SignUpSchema = z.object({
    email: z.string().email({
        message: "Please enter a valid email address"
    }),
    password: z.string().min(8, {
        message: "Password must be at least 8 characters long"
    }),
    name: z.string().min(1, {
        message: "Name is required"
    })
})

export type SignUpSchema = z.infer<typeof SignUpSchema>
export type SignInSchema = z.infer<typeof SignInSchema>